variant context converter refactor
